+++ api/src/main/java/com/beco/api/model/Address.java
@@ -11,7 +11,7 @@ public class Address {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "address_id")
-    private Long addressId;
+    private Integer addressId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "city_id", nullable = false)
@@ -19,5 +19,4 @@ public class Address {
 
     @Column(name = "street", nullable = false)
     private String street;
-
 }
